Firmware snapshot for npu-1.35.0 drop
